Extract expressions into their own file
